@uppy/companion: make validateURL reusable
